app/requirements.txt
asgiref==3.2.10
autopep8==1.5.4
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
cryptography==3.3.2
defusedxml==0.6.0
Django==3.1.13
django-allauth==0.44.0
django-crispy-forms==1.9.2
django-simple-history==2.12.0
django-userforeignkey==0.3.0
idna==2.10
mysqlclient==2.0.1
oauthlib==3.1.0
Pillow==8.3.2
pycodestyle==2.6.0
pycparser==2.20
PyJWT==2.0.0
python3-openid==3.2.0
pytz==2020.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.4.2
toml==0.10.2
urllib3==1.26.5
gunicorn==20.1.0
